Move AST helpers into helpers/ast directory
