+++ src/helpers/ast/to-function-call.js
@@ -1,0 +1,6 @@
+import toReference from "./to-reference";
+
+// Helper to create a function call in AST.
+export default function toFunctionCall(t, functionName, args) {
+  return t.callExpression(toReference(t, functionName), args);
+}
